// src/app/models/itemsdetails.ts
export class Itemsdetails {
    invoiceDetailsId: number;
    invoiceId: number;
    itemid: number;
    itemname: string;
    itemdescription: string;
    serviceOrGoods: number;
    uomId: number;
    uomName: string;
    qty: number;
    taxServiceDesc: string;
    taxvalue: number;
    itemrate: number;
    hsnorsaccode: string;
    taxId: number;
    cgstrate: number;
    cgstvalue: number;
    sgstrate: number;
    sgstvalue: number;
    igstrate: number;
    igstvalue: number;
    total: number;
    currency: number;
    cessrate: number;
    cessvalue: number;
    tdsrate: number;
    tdsvalue: number;
    totalvalue: number;
    servicevalue:number;
    
}